+++ hibernate-core/src/main/java/org/hibernate/boot/internal/AnyKeyType.java
@@ -1,10 +1,10 @@
 /*
  * Hibernate, Relational Persistence for Idiomatic Java
  *
- * SPDX-License-Identifier: Apache-2.0
- * Copyright: Red Hat Inc. and Hibernate Authors
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
+ * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html.
  */
-package org.hibernate.boot.models.bind.internal;
+package org.hibernate.boot.internal;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
