+++ hibernate-core/src/test/java/org/hibernate/orm/test/boot/models/MyStringConverter.java
@@ -4,9 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html.
  */
-package org.hibernate.orm.test.boot.models.process;
-
-import java.util.UUID;
+package org.hibernate.orm.test.boot.models;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
@@ -14,15 +12,15 @@ import jakarta.persistence.Converter;
 /**
  * @author Steve Ebersole
  */
-@Converter
-public class MyUuidConverter implements AttributeConverter<UUID, String> {
+@Converter(autoApply = true)
+public class MyStringConverter implements AttributeConverter<String, String> {
 	@Override
-	public String convertToDatabaseColumn(UUID attribute) {
+	public String convertToDatabaseColumn(String attribute) {
 		return null;
 	}
 
 	@Override
-	public UUID convertToEntityAttribute(String dbData) {
+	public String convertToEntityAttribute(String dbData) {
 		return null;
 	}
 }
