+++ hibernate-core/src/test/java/org/hibernate/orm/test/boot/models/MyUuidConverter.java
@@ -4,7 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html.
  */
-package org.hibernate.orm.test.boot.models.process;
+package org.hibernate.orm.test.boot.models;
+
+import java.util.UUID;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
@@ -12,15 +14,15 @@ import jakarta.persistence.Converter;
 /**
  * @author Steve Ebersole
  */
-@Converter(autoApply = true)
-public class MyStringConverter implements AttributeConverter<String, String> {
+@Converter
+public class MyUuidConverter implements AttributeConverter<UUID, String> {
 	@Override
-	public String convertToDatabaseColumn(String attribute) {
+	public String convertToDatabaseColumn(UUID attribute) {
 		return null;
 	}
 
 	@Override
-	public String convertToEntityAttribute(String dbData) {
+	public UUID convertToEntityAttribute(String dbData) {
 		return null;
 	}
 }
