+++ hibernate-core/src/test/java/org/hibernate/orm/test/boot/models/process/ManagedResourcesSmokeTests.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html.
  */
-package org.hibernate.orm.test.boot.models.resources;
+package org.hibernate.orm.test.boot.models.process;
 
 import java.util.Iterator;
 import java.util.Set;
@@ -50,7 +50,7 @@ import static org.assertj.core.api.Assertions.assertThat;
  * set of "managed classes" (JPA's term) which Hibernate will process for annotations -
  * {@code @Entity}, {@code @Converter}, ...
  * <p/>
- * And is the likely place that hibernate-models-source would hook in to.  Let's see how that might work...
+ * Let's see how using that with hibernate-models that might work...
  *
  * @author Steve Ebersole
  */
